renamed and fixed command (change constants for test)
